core: [Enh] Manage bad formatting of AWS answers to requests
Requests for compil sometimes return '<song title> - <artist>' instead of
clearly identifying them using XML keys. A check box are added to split
this kind of requests. The other check is here to swap the tokens, given
the fact that we get sometimes swapped title and artist.
Model and view were updated to manage that.